Move DataPointer and OwnedPointer to ptr module
Move the safety documenation to module-level docs.
