+++ spatialos-sdk/src/worker/schema.rs
@@ -10,6 +10,7 @@ mod component_data;
 mod component_update;
 mod object;
 mod primitives;
+mod ptr;
 
 pub mod owned;
 
@@ -18,7 +19,7 @@ pub use self::{
     owned::Owned, primitives::*,
 };
 
-pub(crate) use self::private::*;
+pub(crate) use self::ptr::*;
 
 pub type FieldId = u32;
 
@@ -38,87 +39,5 @@ pub trait SchemaPrimitiveField {
             result.push(Self::index(object, field, index));
         }
         result
-    }
-}
-
-// NOTE: We define these traits within a private module so that we can also use them
-// to seal public traits (e.g. `owned::Ownable`).
-mod private {
-    /// A type that acts as an alias to some schema data hidden behind a pointer.
-    ///
-    /// `Schema_Object` and the related schema data "wrapper types" are all represented
-    /// as opaque types hidden behind pointers in the C API; You always work with a
-    /// pointer to the struct, never an instance of the struct itself. To represent
-    /// these in Rust, we define dummy types that can similarly always be put behind a
-    /// reference without ever allowing the user to directly "hold" an instance of the
-    /// type. This allows us to simply cast the raw pointer to a reference to our dummy
-    /// type and preserve the desired semantics of working with a reference, including
-    /// the ownership rules and lifetimes that come with it.
-    ///
-    /// This trait is implemented for all such types, and provides the common methods
-    /// used for type punning with the raw pointer.
-    ///
-    /// # Safety
-    ///
-    /// In order for the implementation of this trait to be safe, the following
-    /// invariants must be upheld:
-    ///
-    /// * It must be a [zero sized type][zst].
-    /// * It must have a private field, such that the type can never be directly
-    ///   instantiated by a user.
-    /// * A reference to `Self` must never be de-referenced. It may only ever be
-    ///   converted to a pointer with `as_ptr()`, which may then be passed to the
-    ///   appropriate functions in the C API.
-    ///
-    /// The `pointer_type_tests!` macro should be used to generate tests for any type
-    /// implementing this trait.
-    ///
-    /// [zst]:  https://doc.rust-lang.org/nomicon/exotic-sizes.html#zero-sized-types-zsts
-    ///
-    /// # Ownable Types
-    ///
-    /// This trait broadly covers all types that can be converted from a pointer to a
-    /// reference, but some of the types (well, all of them except `SchemaObject`) can
-    /// also be owned in addition to being borrowed. For these, see the `owned` module
-    /// and its corresponding `Ownable` trait which extends `DataPointer` with
-    /// additional functionality for creating and destroying owned instances of the
-    /// type.
-    ///
-    /// # Extern Types
-    ///
-    /// There is an [accepted RFC][rfc] for more accurately representing types like
-    /// this. If it is ever implemented and stabilized, we should switch to representing
-    /// these types as extern types, as it would better enforce necessary safety
-    /// invariants.
-    ///
-    /// [rfc]: https://github.com/rust-lang/rfcs/blob/master/text/1861-extern-types.md
-    pub unsafe trait DataPointer: Sized {
-        type Raw;
-
-        unsafe fn from_raw<'a>(raw: *const Self::Raw) -> &'a Self {
-            &*(raw as *const _)
-        }
-
-        unsafe fn from_raw_mut<'a>(raw: *mut Self::Raw) -> &'a mut Self {
-            &mut *(raw as *mut _)
-        }
-
-        fn as_ptr(&self) -> *const Self::Raw {
-            self as *const _ as *const _
-        }
-
-        fn as_ptr_mut(&mut self) -> *mut Self::Raw {
-            self as *mut _ as *mut _
-        }
-    }
-
-    /// A data pointer type that can be owned directly by user code.
-    ///
-    /// All of the data pointer types except `SchemaObject` can be directly owned by the
-    /// user via the `Owned<T>` smart pointer type. This trait defines the constructor
-    /// and destructor functions for the underlying C data needed for this purpose.
-    pub unsafe trait OwnedPointer: DataPointer {
-        const CREATE_FN: unsafe extern "C" fn() -> *mut Self::Raw;
-        const DESTROY_FN: unsafe extern "C" fn(*mut Self::Raw);
     }
 }
